HUGE PROGRESS! See description.
Implemented images for the items. Completely revamped colors, it's dark mode now. Added a "expanded view" for items that will be used to add the funko pops to a cart. Added price fields to each funko pop. Improved display and clickability on each item on main page. Next is to work on the filters and the cart itself.